+++ 05_figures/fig1/phenotype_differences.R
@@ -3,7 +3,7 @@
 #' 
 library(tidyverse)
 
-source("03_scripts/1001genomes_data.R")
+source("03_data_preparation/1001genomes_data.R")
 
 phenotype_diffs <- list(
   audps =